chore(clippy): auto-fix 17 + manual 12 → -D warnings clean on core/ann/mysql
Auto-fixed (cargo clippy --fix):
- useless_conversion, needless_borrow, or_default, needless_borrows_for_generic_args
- unused_mut, useless_conversion in snap/federate/db/acl/server/main/cli

Manual fixes:
- shard.rs: max().min() → .clamp()
- rewrite.rs: 6 regex_creation_in_loops hoisted outside loop
- license/lib.rs: strip_prefix manual → .strip_prefix()
- msql-srv-patched: unused import std::iter removed
- engine/abi.rs: Safety section added, empty doc lines fixed
- engine/rrf.rs: empty line after doc comment fixed
- synapse-ann: doc indent + type_complexity allow
- synapse-core: transmute allow, dead_code allow, type_complexity allows
- synapse-mysql: dead_code allow (bin), type_complexity allows

+++ crates/synapsed/src/main.rs
@@ -422,7 +422,7 @@ async fn search(
     let hits = store.search(q, mode, emb.as_deref(), limit)?;
     let latency_us = t0.elapsed().as_micros() as u64;
     let hit_count = hits.len();
-    let top_score = hits.first().map(|h| h.score as f64).unwrap_or(0.0);
+    let top_score = hits.first().map(|h| h.score).unwrap_or(0.0);
     if let Err(e) = store.log_query(q, mode, latency_us, hit_count, top_score) {
         warn!("query_log insert failed (non-fatal): {e}");
     }
